shifting worker's thread Logic in Presenter class
